fix(CreatibutorsFileModal): Simplify components and CSS rendering
